requirements.txt
vader-sentiment==3.2.1
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.6.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
email-validator==2.2.0
